Updated naming for folders
Add friendly HTML comment to chat.html

A new HTML comment was added at the end of the `chat.html` file. The comment includes a friendly message: "If you see this, then hi!" and is signed by "Olav". This comment does not affect the functionality of the webpage as it is purely informational and intended for developers or anyone viewing the source code.