Bump Version
Bumps the css version that was forgotten in the previous push. This fixes the css not having been loaded because it did not know to reload the css earlier. This method of fixing works by forcing the device to reload the css by changing a slight irrelevant detail in the link tag so that it can force all browsers to reload the latest css file.